requirements.txt
pandas==2.2.3
pydantic==2.9.2
pytest==8.3.3
python-dotenv==1.0.1
responses==0.25.3
requests==2.32.3
